Reduce page refresh timer
Delayed shutdown time is ~10 seconds and new process startup time is either ~<5 or ~>20 seconds depending on whether the HTTPServer's API whitelist is being populated or not. It can take ~60 seconds from the time the client is ready until the HttpServer begins listening again. Allowing for the 20 second grace period from the slowest possible time for a total of 80 seconds is currently acceptable.