deferUntilReady tweak
This should make it possible for this function to be called in userland code while still providing some reasonable debugging information by taking advantage of PHP's stack trace to get the name of the function that called it originally. I don't suppose this library should use it internally, but I'm interested to see what will happen if called from the context of an anonymous function.